Use switch component
