+++ frontend/src/components/index.js
@@ -29,5 +29,6 @@ export { default as TracePreview } from './TracePreview/TracePreview'
 export { default as Preference } from './Preference/Preference'
 export { default as TraceStepBubble } from './TraceStepBubble/TraceStepBubble'
 export { default as Input } from './Input/Input'
+export { default as Switch } from './Switch/Switch'
 export { default as Footer } from './Footer/Footer'
 export { default as ProjectCard } from './ProjectCard/ProjectCard'
